dnsbackend/requirements.txt
Werkzeug==0.14.1
requests==2.18.4
Flask==0.12.2
Flask_Script==2.0.6
SQLAlchemy==1.2.1
xlrd==1.1.0
Flask_SQLAlchemy==2.3.2
Flask_Cors==3.0.3
simplejson==3.16.0
PyJWT==1.6.4
redis==2.10.6
PyMySQL==0.9.2
